[image] Move Copy and CopyFromGzipArchive from partition package
These functions initially were designed to copy data into partition from raw
file as well as from gzip file. This functionality may be used to copy file to
file and gzip to file. That's why it is moved into image package. Additional
parameter createDst is added. This parameter indicates shall destination file
be created if not exist or not. For copying partition this parameter shall be
set to false.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>