[resourcemonitor] Add update iptables filter cache
Instead of making a request to iptables with each call to
GetInstanceTraffic and GetSystemTraffic that blocked
with option --wait or blocked the /var/run/xtables.lock file,
the functionality was added to get all iptables rules
and cache them, and now, when calling the GetInstanceTraffic
and GetSystemTraffic functions, they took metrics from the cache
instead of a direct request to iptables. This made it possible
to reduce blocking when a network is created for an instance
and at the same time network monitoring takes place.

Signed-off-by: Mykola Solianko <mykola_solianko@epam.com>